Draw outTube, added border and color for tubes
I found out why outline wasn't showing up, if thickness is not specified, the default is 0 pixels. I also added colours to the tubes. Do you think the balls should be hidden while in the tube or should they visible to give the player some advantage? Or it could be a feature where it's visible only when the player's mouse is hovering over it.