Renaming variables
I don't think we should be setting the balls' position and drawing everything based around that. I might want to add more info to the level class about which ball comes out of where if we have a level (probably level 3) with multiple inTubes. Since right now everything is uniform with starting ball positions it's only renaming variables at this point which would be the best time to do it.