checkForSQSelect() and debugger statements
Checks if a stack or queue has been clicked on and triggers the appropriate tStack method depending on whether it was a right or left mouse button click. Left button for pop() and right button for top(). Also modifies the Ball class boolean flags, isPushed and isTopped accordingly. Debugger statements are for next and prev ball to keep an eye on the E and Q bug.