LevelScreen, Lock/Unlock levels and debugger statements.
This version has a bug where level 1 is automatically won, if the
reinitialise statements are placed in the “if(!gameRunning)” section of
the main program’s while loop in main.cpp, level one plays no matter
which level is chosen. Been placing debugger statements, modifying code
and staring at this for 2 days now.