Level Selection System but double count bug
The level selection system works fine, this version uses a single
global array to check the list of locked levels and remove them
appropriately using new functions in main.cpp. The same array is used
by LevelScreen.cpp to check if a lock needs to be drawn over a level
button or not. Using a royalty free clip-art png for the lock.

The unlocked attribute of a level is not used, will remove later if no
use is found.

There is a double count bug with the Q and E keys and this has been
reported to issues a while ago. The quick fix was swapping the keys but
that doesn’t work for a number of balls different than 3.