Help page
Help page is accessible from a button on the main menu screen. Its
design is pretty basic though. Laid down the foundation for an overhaul
in the menu system design where all non-gameplay related screens are
handled by a single of pair of .cpp and .hpp pages. So far the help
page and level select page (type 2 and 3 respectively) have been
implemented this way. Main menu will be type 1.